# packages/spec/swagger.json
{"openapi":"3.0.0","paths":{"/info":{"get":{"operationId":"getInfo","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppInfoViewModel"}}}}},"tags":["App"]}},"/audits/list":{"post":{"operationId":"listAudits","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditListViewModel"}}}}},"tags":["audits"]}},"/jobs/applications/list":{"post":{"operationId":"listApplications","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiPaginatedRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationListViewModel"}}}}},"tags":["applications"]}},"/jobs/applications/upload":{"post":{"operationId":"uploadApplications","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/ApplicationsUploadViewModel"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationViewModel"}}}}},"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["applications"]}},"/jobs/applications/bulk":{"post":{"operationId":"createApplications","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationViewModel"}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["applications"]}},"/jobs/applications":{"post":{"operationId":"createApplication","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationViewModel"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["applications"]}},"/jobs/applications/{applicationId}":{"put":{"operationId":"updateApplication","parameters":[{"name":"applicationId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationViewModel"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationViewModel"}}}}},"tags":["applications"]},"get":{"operationId":"getApplication","parameters":[{"name":"applicationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationViewModel"}}}}},"tags":["applications"]},"delete":{"operationId":"deleteApplication","parameters":[{"name":"applicationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["applications"]}},"/jobs/companies":{"get":{"operationId":"getCompanies","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyListViewModel"}}}}},"tags":["companies"]}},"/jobs/companies/list":{"post":{"operationId":"getCompaniesList","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyFullListViewModel"}}}}},"tags":["companies"]}},"/jobs/companies/{companyId}":{"delete":{"operationId":"deleteCompany","parameters":[{"name":"companyId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["companies"]}}},"info":{"title":"API","description":"The main API for the UI","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"AppInfoViewModel":{"type":"object","properties":{"version":{"type":"string"}},"required":["version"]},"EnumAuditActionTypes":{"type":"string","enum":["Created","Updated","Deleted"],"x-enum-varnames":["created","updated","deleted"]},"AuditDiffViewModel":{"type":"object","properties":{"action":{"allOf":[{"$ref":"#/components/schemas/EnumAuditActionTypes"}]},"field":{"type":"string"},"valueCurrent":{"type":"object"},"valuePrevious":{"type":"object"}},"required":["action","field"]},"AuditViewModel":{"type":"object","properties":{"action":{"allOf":[{"$ref":"#/components/schemas/EnumAuditActionTypes"}]},"id":{"type":"string"},"userId":{"type":"string"},"entity":{"type":"string"},"diff":{"type":"array","items":{"$ref":"#/components/schemas/AuditDiffViewModel"}},"dateCreated":{"type":"number"}},"required":["action","id","userId","entity","diff","dateCreated"]},"AuditListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AuditViewModel"}},"total":{"type":"number"}},"required":["data"]},"ApiPaginatedRequest":{"type":"object","properties":{"start":{"type":"number"},"limit":{"type":"number"},"page":{"type":"number"}},"required":["start","limit","page"]},"EnumApplicationStatus":{"type":"number","enum":[-1,0,1,2,3,4,5,6,7],"x-enum-varnames":["Applied","CurrentWeek","Initial","Interviewing","InterviewedAndRejected","Rejected","Declined","Accepted","Ghosted"]},"CompanyViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"id":{"type":"string"},"name":{"type":"string"}},"required":["id","name"]},"CommentViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"id":{"type":"string"},"comment":{"type":"string"},"applicationId":{"type":"string"}},"required":["id","comment","applicationId"]},"ApplicationViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"status":{"allOf":[{"$ref":"#/components/schemas/EnumApplicationStatus"}]},"id":{"type":"string"},"positionTitle":{"type":"string"},"dateApplied":{"type":"number"},"url":{"type":"string"},"site":{"type":"string"},"compensation":{"type":"string"},"company":{"$ref":"#/components/schemas/CompanyViewModel"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentViewModel"}}},"required":["status","id","positionTitle","dateApplied","url","site","compensation","company","comments"]},"ApplicationListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationViewModel"}},"total":{"type":"number"}},"required":["data"]},"ApplicationsUploadViewModel":{"type":"object","properties":{"addHeaders":{"type":"string","format":"boolean"},"file":{"type":"string","format":"binary"}},"required":["addHeaders","file"]},"CompanyListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CompanyViewModel"}},"total":{"type":"number"}},"required":["data"]},"ApplicationNestedViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"status":{"allOf":[{"$ref":"#/components/schemas/EnumApplicationStatus"}]},"id":{"type":"string"},"positionTitle":{"type":"string"},"dateApplied":{"type":"number"},"url":{"type":"string"},"site":{"type":"string"},"compensation":{"type":"string"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentViewModel"}}},"required":["status","id","positionTitle","dateApplied","url","site","compensation","comments"]},"CompanyFullViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"id":{"type":"string"},"name":{"type":"string"},"applications":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationNestedViewModel"}}},"required":["id","name","applications"]},"CompanyFullListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CompanyFullViewModel"}},"total":{"type":"number"}},"required":["data"]}}}}